requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aeon=0.4.0=pyhd8ed1ab_0
asttokens=2.4.0=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
brotli=1.0.9=h9c3ff4c_4
brotli-python=1.0.9=py310hd8f1fbe_9
bzip2=1.0.8=h7f98852_4
c-ares=1.19.1=hd590300_0
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py310h255011f_3
charset-normalizer=3.2.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
conda-package-handling=2.0.2=pyh38be061_0
conda-package-streaming=0.8.0=pyhd8ed1ab_0
contourpy=1.1.1=py310hd41b1e2_1
cryptography=41.0.2=py310h75e40e8_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.8.0=py310hc6cd4ac_1
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pyh1a96a4e_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fmt=9.1.0=h924138e_0
fonttools=4.43.0=py310h2372a71_0
freetype=2.12.1=h267a509_2
gettext=0.21.1=h27087fc_0
icu=72.1=hcb278e6_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
importlib_resources=6.1.0=pyhd8ed1ab_0
ipykernel=6.25.0=pyh71e2992_0
ipython=8.16.0=pyh0d859eb_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py310h06a4308_0
joblib=1.3.0=pyhd8ed1ab_1
jsonschema=4.19.1=pyhd8ed1ab_0
jsonschema-specifications=2023.7.1=pyhd8ed1ab_0
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=5.3.2=py310hff52083_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py310hd41b1e2_1
krb5=1.21.1=h659d440_0
lame=3.100=h166bdaf_1003
lcms2=2.15=haa2dc70_1
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libarchive=3.6.2=h039dbb9_1
libblas=3.9.0=18_linux64_openblas
libcblas=3.9.0=18_linux64_openblas
libcurl=8.1.2=hca28451_1
libdeflate=1.18=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.4.2=h7f98852_5
libflac=1.4.3=h59595ed_0
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=13.1.0=he5830b7_0
libiconv=1.17=h166bdaf_0
libjpeg-turbo=2.1.5.1=hd590300_1
liblapack=3.9.0=18_linux64_openblas
libllvm14=14.0.6=hcd5def8_4
libmamba=1.4.2=hcea66bb_0
libmambapy=1.4.2=py310h1428755_0
libnghttp2=1.52.0=h61bc06f_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.24=pthreads_h413a1c8_0
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h753d276_0
libsndfile=1.2.2=hc60ed4a_1
libsodium=1.0.18=h36c2ea0_1
libsolv=0.7.24=hfc55251_1
libsqlite=3.42.0=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtiff=4.5.1=h8b53f26_1
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.15=h0b41bf4_0
libxml2=2.11.4=h0d562d8_0
libzlib=1.2.13=hd590300_5
llvmlite=0.40.1=py310h1b8f574_0
lz4-c=1.9.4=hcb278e6_0
lzo=2.10=h516909a_1000
markupsafe=2.1.3=py310h2372a71_1
matplotlib-base=3.7.2=py310hf38f957_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpg123=1.32.2=h59595ed_0
munkres=1.1.4=pyh9f0ad1d_0
nbformat=5.9.1=pyhd8ed1ab_0
ncurses=6.4=hcb278e6_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pyhd8ed1ab_0
numba=0.57.1=py310h0f6aa51_0
numpy=1.24.4=py310ha4c1d20_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.1.4=hd590300_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.0.3=py310h7cbd5c2_1
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.0=py310h582fbeb_0
pip=23.1.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=3.10.0=pyhd8ed1ab_0
plotly=5.15.0=pyhd8ed1ab_0
pluggy=1.2.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=py310h2372a71_1
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pybind11-abi=4=hd8ed1ab_3
pycosat=0.6.4=py310h5764c6d_1
pycparser=2.21=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
pysoundfile=0.12.1=pyhd8ed1ab_0
python=3.10.12=hd12c33a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.18.1=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytz=2023.3.post1=pyhd8ed1ab_0
pyzmq=25.1.1=py310h5bbb5d0_1
readline=8.2=h8228510_1
referencing=0.30.2=pyhd8ed1ab_0
reproc=14.2.4=h0b41bf4_0
reproc-cpp=14.2.4=hcb278e6_0
requests=2.31.0=pyhd8ed1ab_0
rpds-py=0.10.3=py310hcb5633a_1
ruamel.yaml=0.17.32=py310h2372a71_0
ruamel.yaml.clib=0.2.7=py310h1fa729e_1
scikit-learn=1.2.2=py310hf7d194e_2
scipy=1.11.3=py310hb13e2d6_0
seaborn=0.13.0=hd8ed1ab_0
seaborn-base=0.13.0=pyhd8ed1ab_0
setuptools=68.0.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=py310h278f3c1_1
tenacity=8.2.3=pyhd8ed1ab_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.12=h27826a3_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.3.3=py310h2372a71_1
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.10.1=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
unicodedata2=15.1.0=py310h2372a71_0
urllib3=2.0.3=pyhd8ed1ab_1
wcwidth=0.2.8=pyhd8ed1ab_0
wfdb=4.1.2=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
wrapt=1.15.0=py310h2372a71_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
yaml-cpp=0.7.0=h27087fc_2
zeromq=4.3.4=h9c3ff4c_1
zipp=3.17.0=pyhd8ed1ab_0
zstandard=0.19.0=py310h1275a96_2
zstd=1.5.2=hfc55251_7
